alex/default-to-svg: move out of folder
