+++ packages/editor/src/lib/exports/embedMedia.ts
